main: Fix class name and package restriction errors
