+++ Java/PyramidWithStarsPattern.java
@@ -1,17 +1,18 @@
+package Java;
+
 public class PyramidWithStarsPattern {
 
 	// function to print the following pyramid
 	// pattern
-	static void printPattern(int n)
-	{
+	static void printPattern(int n) {
 		int j, k = 0;
 
 		// loop to decide the row number
 		for (int i = 1; i <= n; i++) {
-			
+
 			// if row number is odd
 			if (i % 2 != 0) {
-			
+
 				// print numbers with the '*'
 				// sign in increasing order
 				for (j = k + 1; j < k + i; j++)
@@ -24,7 +25,7 @@ public class PyramidWithStarsPattern {
 
 			// if row number is even
 			else {
-				
+
 				// update value of 'k'
 				k = k + i - 1;
 
@@ -38,9 +39,8 @@ public class PyramidWithStarsPattern {
 	}
 
 	// Driver program to test above
-public static void main(String args[])
-	{
-		int n =4;
+	public static void main(String[] args) {
+		int n = 4;
 		printPattern(n);
 	}
 }
